src/domain/is-kanji-number-follewed-by-cho.ts
/*!
 * MIT License
 *
 * Copyright (c) 2023 デジタル庁
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */
import { findKanjiNumbers } from '@geolonia/japanese-numeral';
import { RegExpEx } from './reg-exp-ex';

/**
 * targetTownName 内に 「漢数字 + 町」の地名がある場合、trueを返す
 * @param targetTownName
 * @returns
 */
export const isKanjiNumberFollewedByCho = (targetTownName: string) => {
  const xCho = targetTownName.match(RegExpEx.create('.町', 'g'));
  if (!xCho) return false;

  const kanjiNumbers = findKanjiNumbers(xCho[0]);
  return kanjiNumbers.length > 0;
};
